Added summoning sickness to the gates and shepherds.

+++ src/main/scala/org/frankversnel/poortjes/game/gameobjects/player_killer.scala
@@ -3,9 +3,11 @@ package org.frankversnel.poortjes.game.gameobjects
 import org.frankversnel.poortjes.GameObject
 import org.frankversnel.poortjes.collision.Collidable
 
-trait PlayerKiller extends Collidable {
+trait PlayerKiller extends Collidable with SummoningSickness {
+	protected val maxSicknessDurationMillis = 3000L
+
 	def onCollision(collider: GameObject) {
-		if(collider.is[Player]) {
+		if(collider.is[Player] && !hasSummoningSickness) {
 			collider.destroy
 		}
 	}
